decouple the strong relationship between UserService and AuthenticationProvider by introducing a UserCreationService
also reduce the use of @Autowired
