+++ src/main/java/de/tum/in/www1/artemis/service/connectors/AbstractVersionControlService.java
@@ -11,7 +11,6 @@ import org.apache.commons.io.FileUtils;
 import org.eclipse.jgit.api.errors.GitAPIException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.ApplicationContext;
 
@@ -36,20 +35,16 @@ public abstract class AbstractVersionControlService implements VersionControlSer
     @Value("${artemis.lti.user-prefix-u4i:#{null}}")
     protected Optional<String> userPrefixU4I;
 
-    private ApplicationContext applicationContext;
+    private final ApplicationContext applicationContext;
 
     protected final UrlService urlService;
 
     private final GitService gitService;
 
-    public AbstractVersionControlService(UrlService urlService, GitService gitService) {
+    public AbstractVersionControlService(ApplicationContext applicationContext, UrlService urlService, GitService gitService) {
+        this.applicationContext = applicationContext;
         this.urlService = urlService;
         this.gitService = gitService;
-    }
-
-    @Autowired
-    public void setApplicationContext(ApplicationContext applicationContext) {
-        this.applicationContext = applicationContext;
     }
 
     /**
