+++ src/main/java/de/tum/in/www1/artemis/web/rest/CustomMetricsExtension.java
@@ -3,7 +3,6 @@ package de.tum.in.www1.artemis.web.rest;
 import java.util.HashMap;
 import java.util.Map;
 
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.actuate.endpoint.annotation.ReadOperation;
 import org.springframework.boot.actuate.endpoint.web.annotation.EndpointWebExtension;
 import org.springframework.messaging.simp.user.SimpUserRegistry;
@@ -19,10 +18,8 @@ import io.github.jhipster.config.metric.JHipsterMetricsEndpoint;
 @EndpointWebExtension(endpoint = JHipsterMetricsEndpoint.class)
 public class CustomMetricsExtension {
 
-    @Autowired
     private final JHipsterMetricsEndpoint jHipsterMetricsEndpoint;
 
-    @Autowired
     private final SimpUserRegistry simpUserRegistry;
 
     public CustomMetricsExtension(JHipsterMetricsEndpoint jHipsterMetricsEndpoint, SimpUserRegistry simpUserRegistry) {
