+++ src/test/java/de/tum/in/www1/artemis/service/connectors/LtiServiceTest.java
@@ -27,13 +27,13 @@ import de.tum.in.www1.artemis.repository.LtiUserIdRepository;
 import de.tum.in.www1.artemis.repository.ResultRepository;
 import de.tum.in.www1.artemis.repository.UserRepository;
 import de.tum.in.www1.artemis.security.ArtemisAuthenticationProvider;
-import de.tum.in.www1.artemis.service.user.UserService;
+import de.tum.in.www1.artemis.service.user.UserCreationService;
 import de.tum.in.www1.artemis.web.rest.dto.LtiLaunchRequestDTO;
 
 public class LtiServiceTest {
 
     @Mock
-    private UserService userService;
+    private UserCreationService userCreationService;
 
     @Mock
     private UserRepository userRepository;
@@ -71,7 +71,7 @@ public class LtiServiceTest {
     public void init() {
         MockitoAnnotations.openMocks(this);
         SecurityContextHolder.clearContext();
-        ltiService = new LtiService(userService, userRepository, ltiOutcomeUrlRepository, resultRepository, artemisAuthenticationProvider, ltiUserIdRepository, response);
+        ltiService = new LtiService(userCreationService, userRepository, ltiOutcomeUrlRepository, resultRepository, artemisAuthenticationProvider, ltiUserIdRepository, response);
         Course course = new Course();
         course.setStudentGroupName(courseStudentGroupName);
         exercise = new TextExercise();
@@ -152,7 +152,7 @@ public class LtiServiceTest {
         user.setActivated(false);
         when(ltiUserIdRepository.findByLtiUserId(launchRequest.getUser_id())).thenReturn(Optional.empty());
         when(userRepository.findOneByLogin(username)).thenReturn(Optional.empty());
-        when(userService.createUser(username, groups, "", launchRequest.getLis_person_sourcedid(), launchRequest.getLis_person_contact_email_primary(), null, null, "en"))
+        when(userCreationService.createUser(username, groups, "", launchRequest.getLis_person_sourcedid(), launchRequest.getLis_person_contact_email_primary(), null, null, "en"))
                 .thenReturn(user);
 
         onSuccessfulAuthenticationSetup(user, ltiUserId);
@@ -160,7 +160,7 @@ public class LtiServiceTest {
         ltiService.handleLaunchRequest(launchRequest, exercise);
 
         onSuccessfulAuthenticationAssertions(user, ltiUserId);
-        verify(userService).activateUser(user);
+        verify(userCreationService).activateUser(user);
 
         SecurityContextHolder.clearContext();
         launchRequest.setContext_label("randomLabel");
@@ -216,7 +216,7 @@ public class LtiServiceTest {
         assertThat("ff30145d6884eeb2c1cef50298939383".equals(ltiUserId.getLtiUserId()));
         assertThat("some.outcome.service.url.com".equals(ltiOutcomeUrl.getUrl()));
         assertThat("someResultSourceId".equals(ltiOutcomeUrl.getSourcedId()));
-        verify(userService, times(1)).saveUser(user);
+        verify(userCreationService, times(1)).saveUser(user);
         verify(artemisAuthenticationProvider, times(1)).addUserToGroup(user, courseStudentGroupName);
         verify(ltiOutcomeUrlRepository, times(1)).save(ltiOutcomeUrl);
     }
